[cast] Move cast to util/cast, it's too small
- cast was based on https://github.com/spf13/cast but it only
implemented a small amount of methods because in config we no longer use
string path to find value in a complex untyped struct, and unmarshal it
to struct directly
