+++ directory.md
@@ -1,0 +1,22 @@
+# Directory
+
+## Project Layout
+
+- [cmd/gommon](cmd/gommon) the command line application
+- [config](config) config file reader
+- [doc](doc) documentation
+- [errors](errors) error wrapping and multi error
+- [generator](generator) generating interface methods, render go template, protobuf etc.
+- [legacy](legacy) legacy code base
+- [noodle](noodle) embed static assets for go binary with .ignore file support
+- [playground](playground) test library and replay issues
+- [requests](requests) ~~http util~~
+- [scripts](scripts) test scripts
+- [structure](structure) data structure
+- [util](util) small packages
+
+## Package Layout
+
+- README.md 
+- doc
+  - history.md verbose log about the evolve of the package, important decision made and the reason behind it
