+++ doc/README.md
@@ -1,2 +1,7 @@
 # Gommon Documentation
 
+- Style
+  - [General](style.md)
+  - [Application using Gommon](style-application.md)
+  - [Library using Gommon](style-library.md)
+  - [Writing Gommon](style-gommon.md)
