--- doc/directory.md
@@ -1,15 +1,0 @@
-# Directory
-
-- [cmd/gommon](../cmd/gommon) the command line application
-- [cast](../cast) converter needed for config
-- [config](../config) config file reader
-- [doc](.) documentation
-- [errors](../errors) error wrapping and multi error
-- [generator](../generator) generating interface methods, render go template, protobuf etc.
-- [legacy](../legacy) legacy code base
-- [noodle](../noodle) embed static assets for go binary with .ignore file support
-- [playground](../playground) test library and replay issues
-- [requests](../requests) http util
-- [scripts](../scripts) test scripts
-- [structure](../structure) data structure
-- [util](../util) util packages for public use, they are too small to be top level package
