+++ doc/style-gommon.md
@@ -1,0 +1,18 @@
+# Gommon Style
+
+Style for writing gommon itself, see [library](style-library.md) and [application](style-application.md) for using gommon.
+
+## Folder structure
+
+see [directory](../directory.md)
+
+- [ ] TODO: add how to structure file inside a package etc.
+
+## Documentation
+
+MUST cover the following
+
+- [ ] convention, i.e. variable names, error handling etc.
+- [ ] internal, a basic walk through of import parts
+- define canonical import path in `pkg.go` https://golang.org/doc/go1.4#canonicalimports 
+- [ ] survey
