+++ scripts/docker-compose.yml
@@ -2,7 +2,7 @@ version: '2'
 services:
 # FIXME: there should be a way of reusing command, I am just copying stuff around, there is extend keyword, but removed in v3
   golang1.9:
-    image: golang:1.9
+    image: golang:1.10
     command: go test -v -cover ./cast/... ./config/... ./generator/... ./log/... ./playground/... ./requests/... ./structure/... ./util/...
     working_dir: /go/src/github.com/dyweb/gommon
     volumes:
