+++ util/cast/pkg.go
@@ -1,4 +1,4 @@
 // Package cast convert types safely and drop incompatible types during conversion
 //
 // it is inspired by https://github.com/spf13/cast
-package cast // import "github.com/dyweb/gommon/cast"
+package cast // import "github.com/dyweb/gommon/util/cast"
